+++ src/implementation/test/test/implementation/ConcurrentPrimeFinderTest.java
@@ -19,9 +19,9 @@ public class ConcurrentPrimeFinderTest {
     public void addEntry() throws Exception {
         // TODO Check
         ValidatingPrimeSet set = new ValidatingPrimeSet();
-        List<Long> primes = Arrays.asList(5L,7L,29L,47L,59L,61L,67L,79L,83L,89L,269L,463L,467L,487L,569L,599L,859L,883L,887L);
+        List<Integer> primes = Arrays.asList(5,7,29,47,59,61,67,79,83,89,269,463,467,487,569,599,859,883,887);
         boolean isValid = true;
-        for (Long prime : primes ) {
+        for (Integer prime : primes ) {
             if (!set.addEntry(prime) ){
                 isValid = false;
                 break;
@@ -48,7 +48,7 @@ public class ConcurrentPrimeFinderTest {
     @Test
     public void primesFrom0To1000_checkCount(){
         ConcurrentPrimeFinder finder = new ConcurrentPrimeFinder();
-        List<Long> primes = finder.findPrimes(0, 1000);
+        List<Integer> primes = finder.findPrimes(0, 1000);
 
         Assert.assertEquals("There should be 168 primes between 0 and 1000", 168, primes.size());
     }
@@ -56,9 +56,9 @@ public class ConcurrentPrimeFinderTest {
     @Test
     public void primesFrom0To100(){
         ConcurrentPrimeFinder finder = new ConcurrentPrimeFinder();
-        List<Long> primes = finder.findPrimes(0, 100);
+        List<Integer> primes = finder.findPrimes(0, 100);
 
-        List<Long> expected = Arrays.asList(2L,3L,5L,7L,11L,13L,17L,19L,23L,29L,31L,37L,41L,43L,47L,53L,59L,61L,67L,71L,73L,79L,83L,89L,97L);
+        List<Integer> expected = Arrays.asList(2,3,5,7,11,13,17,19,23,29,31,37,41,43,47,53,59,61,67,71,73,79,83,89,97);
 
         //Assert.assertEquals("There should be 25 primes between 0 and 100", 25, primes.size());
         Assert.assertEquals("All primes should be found", expected, primes);
